% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interrater_Merger.R
\name{Interrater_Merger}
\alias{Interrater_Merger}
\title{A Function to Automatically Merge the Revised Annotations after Interrater Revision.}
\usage{
Interrater_Merger(
  Rater1_Directory,
  R1_Keyword = FALSE,
  R1_Type = ".txt",
  R1_Deliminator = "\\t",
  R1 = "",
  IR_Directory,
  IR_Keyword = FALSE,
  IR_Type = ".txt",
  IR_Deliminator = "\\t",
  IR = "",
  Artifact,
  Artifact_Action = TRUE,
  Uniform_Output = TRUE,
  ...
)
}
\arguments{
\item{Rater1_Directory}{Directory of where the annotations were made by the first rater t <e.g. "C:/Users/__YOUR USERNAME__/UPSTREAM FOLDER/.../FOLDER NAME/">}

\item{R1_Keyword}{Specify keywords for the annotation files. Make sure to provide the maximum patterns shared among files.}

\item{R1_Type}{The type of file to import. Currently, supports  ".tsv. <default: ".txt">}

\item{R1_Deliminator}{If the data is stored in .txt file and the deliminator is not TAB as default, please, specify it. <default: "\t">}

\item{R1}{The initial of rater 1.}

\item{Artifact}{A value serves as artifact labeling}

\item{Artifact_Action}{A logical parameter indicating the action used in case of artifact labeling between raters. If TRUE, artifacts mis-alignments will be treated as differences between raters. If FALSE, the non-artifact label will be used as the annotation for the mini-epoch. When FALSE was passed to the function, it will generate a new column called 'Artifacts'. <Default: "TRUE">}

\item{Uniform_Output}{To specify if the user wish to export files containing interrater differences into the first rater's folder, TRUE, or in duplicates downstream of folders that contain raters annotation separately, FALSE. <Default: TRUE>}

\item{Revised_Directory}{Directory of where the annotations were made by the second rater <e.g. "C:/Users/___YOUR USERNAME___/UPSTREAM FOLDER/.../FOLDER NAME/">}

\item{RA_Keyword}{Specify keywords for the stage files. Make sure to provide the maximum patterns shared among files.}

\item{RA_Type}{The type of file to import. Currently, supports  ".tsv . <default: ".txt">}

\item{RA_Deliminator}{If the data is stored in .txt file and the deliminator is not TAB as default, please, specify it. <default: "\t">}

\item{RA}{The initial of rater 2.}
}
\value{
Export interrater differences for each annotation file for further revision
}
\description{
This function will loop over revised annotations after interrater difference comparisons and merge them with the original annotation. These differences will be extracted and stored in a new folders downstream to the output directory assigned.
*** It is recommended that raters assign their initials following an underscore after
}
\examples{

Interrater_Merger(Rater1_Directory,R1_Type=".txt", R1_Deliminator="\t",R1="",
                  IR_Directory,IR_Type=".txt", IR_Deliminator="\t",IR="")




}
\keyword{Annotation}
\keyword{Interrater}
\keyword{Merge}
\keyword{Merger}
\keyword{interrater}
\keyword{merge}
\keyword{merger}
