% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stage_Ann_Checker.R
\name{Stage_Ann_Checker}
\alias{Stage_Ann_Checker}
\title{A Function to Automatically Check for Annotation Errors}
\usage{
Stage_Ann_Checker(
  Annotation_Directory,
  AF_Keyword = FALSE,
  AF_Type = ".txt",
  AF_Deliminator = "\\t",
  Stage_Directory,
  SF_Keyword = FALSE,
  SF_Type = ".txt",
  SF_Deliminator = "\\t",
  cOnset = "Onset",
  cDuration = "Duration",
  cAnnotation = "Annotation",
  Arousal = "0",
  N1 = "1",
  N2 = "2",
  N3 = "3",
  N4 = "4",
  REM = "5",
  Unscored = "9",
  Stage_Duration = 30,
  Epoch_Duration = 3,
  Miniepoch = "5",
  Target_Keyword = "_W",
  ID_List,
  ...
)
}
\arguments{
\item{Annotation_Directory}{Directory of where the files that you want to import <e.g. "C:/Users/__YOUR USERNAME__/UPSTREAM FOLDER/.../FOLDER NAME/">}

\item{AF_Keyword}{Specify keywords for the annotation files. Make sure to provide the maximum patterns shared among files.}

\item{AF_Type}{The type of file to import. Currently, supports  ".tsv. <default: ".txt">}

\item{AF_Deliminator}{If the data is stored in .txt file and the deliminator is not TAB as default, please, specify it. <default: "\t">}

\item{Stage_Directory}{Directory of where the Stage files are stored  <e.g. "C:/Users/___YOUR USERNAME___/UPSTREAM FOLDER/.../FOLDER NAME/">}

\item{SF_Keyword}{Specify keywords for the stage files. Make sure to provide the maximum patterns shared among files.}

\item{SF_Type}{The type of file to import. Currently, supports  ".tsv . <default: ".txt">}

\item{SF_Deliminator}{If the data is stored in .txt file and the deliminator is not TAB as default, please, specify it. <default: "\t">}

\item{cOnset}{The name of the column that contains the onset time for each epoch in accumulated seconds in both annotation files and stage files <Default: "Onset">}

\item{cDuration}{The name of the column that contains durations for each epoches in both annotation files and stage files <Default: "Duration">}

\item{cAnnotation}{The name of the column that contains annotations for both annotation files and stage files <Default: "Annotation">}

\item{Arousal}{Specify the annotations indicating arousal <Default: "0">}

\item{N1}{Specify the annotations indicating Stage 1 NREM <Default: "1">}

\item{N2}{Specify the annotations indicating Stage 2 NREM <Default: "2">}

\item{N3}{Specify the annotations indicating Stage 3 NREM <Default: "3">}

\item{N4}{Specify the annotations indicating Stage 4 NREM <Default: "4">}

\item{REM}{Specify the annotations indicating Stage REM <Default: "5">}

\item{Unscored}{Specify the annotations indicating Unscored Stage  <Default: "9">}

\item{Stage_Duration}{Specify the duration of each stage  <Default: 30>}

\item{Epoch_Duration}{Specify the duration of the miniepochs  <Default: 3>}

\item{Miniepoch}{Specify which stages contain miniepoches  <Default: "5">}

\item{Target_Keyword}{Specify keywords for the error files  <Default: "_W>}

\item{ID_List}{a list of IDs that contains the file names}
}
\value{
Import files into your working enviroment in R (i.e., Global Enviroment in R Studio)
}
\description{
This function will loop over annotations in a folder checking for missing mini-epochs within a stage and misalignment in sleep stages.
}
